Modify several controllers, adminRoutes and jwt.js

+++ controllers/adminControllers/hardDeleteUser.js
@@ -9,7 +9,11 @@ const hardDeleteUser = async (req, res) => {
       if (!user) {
         return res.status(404).json({ message: "User not found" });
       }
-  
+      
+      if (!user.canDeleteUser(user)) {
+        return res.status(403).json({message: "You cannot delete your admin account."});
+      }
+      
       await user.destroy({ force: true });
   
       await AdminLog.create({
